+++ packages/form-render/src/form-items/index.ts
@@ -4,6 +4,7 @@ import { renderSelectItem } from './select'
 import { renderSwitchItem } from './switch'
 import { renderDateRangeItem } from './date-range'
 import { renderTextareaItem } from './textarea'
+import { renderTreeSelectItem } from './tree-select'
 import { renderRenderItem } from './render'
 
 export function FormItemRenders<T>() {
@@ -11,6 +12,7 @@ export function FormItemRenders<T>() {
     input: renderInputItem<T>,
     textarea: renderTextareaItem<T>,
     select: renderSelectItem<T>,
+    treeSelect: renderTreeSelectItem<T>,
     date: renderDateItem<T>,
     switch: renderSwitchItem<T>,
     dateRange: renderDateRangeItem<T>,